Updates to the functionality of the GUI
Got the slider to correctly read in duration for static mode. Real time mode is working as well. However I've had to cap the max duration for real time at 10 seconds for now because the arrays fill up... which is something that needs some work. Implemented popup messages if radiobuttons aren't selected